backend/requirements.txt
fastapi>=0.123.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-genai>=0.2.0
google-generativeai>=0.8.0
supabase>=2.9.0
pydantic>=2.12.0
python-multipart==0.0.6
stripe==7.0.0
Pillow==10.1.0
requests>=2.31.0

